r/aws_bedrocksgentcore_memory_strategy: Consolidate test files.
